app/requirements.txt
# Name: requirements.txt
# Version: 0.1.0
# Created: 971201
# Modified: 250502
# Creator: ParcoAdmin
# Modified By: ParcoAdmin
# Description: Text file for ParcoRTLS configuration
# Location: /home/parcoadmin/parco_fastapi/app
# Role: Utility
# Status: Active
# Dependent: TRUE

annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dnspython==2.7.0
eventlet==0.39.0
fastapi==0.115.8
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.2
opencv-python-headless==4.11.0.86
packaging==24.2
paho-mqtt==2.1.0
pillow==11.1.0
pluggy==1.5.0
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
pytest-asyncio==0.25.3
python-engineio==4.11.2
python-multipart==0.0.20
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==15.0.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
